Nest ProjectSnapshotManagerDispatcher within ProjectSnapshotManager
